+++ packages/stargate/src/modules/index.ts
@@ -86,3 +86,9 @@ export {
 } from "./staking/messages";
 export { setupStakingExtension, StakingExtension } from "./staking/queries";
 export { setupTxExtension, TxExtension } from "./tx/queries";
+export {
+  AminoMsgCreateVestingAccount,
+  createVestingAminoConverters,
+  isAminoMsgCreateVestingAccount,
+} from "./vesting/aminomessages";
+export { vestingTypes } from "./vesting/messages";