+++ packages/stargate/src/modules/vesting/messages.ts
@@ -1,0 +1,6 @@
+import { GeneratedType } from "@cosmjs/proto-signing";
+import { MsgCreateVestingAccount } from "cosmjs-types/cosmos/vesting/v1beta1/tx";
+
+export const vestingTypes: ReadonlyArray<[string, GeneratedType]> = [
+  ["/cosmos.vesting.v1beta1.MsgCreateVestingAccount", MsgCreateVestingAccount],
+];