+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ### Added
 
 - @cosmjs/encoding: Create `normalizeBech32`.
+- @cosmjs/stargate: Added support for vesting messages.
 
 ## [0.28.1] - 2022-03-30